Update MainFrame.java
Just wanted to add a comment to warn programmers not remove the line of code without good reason. 

Without setting windowless_rendering_enabled=false the sample app browser would lock up with a spinning cursor (windows) or simply no response from clicks (linux). 

But I didn't want to label OSR as experimental or buggy-- (but it might be!?) Just not a clear setting and strange the default causes buggy behavior.